#228 [B] Optimize File Upload and Calculation Speeds
I have drastically increased the speed in which files for student roster
are able to be uplaoded. The calculation speed is also drastically
improved

* Both the Students and Grade models are created with a bulk
  create operation which is drastically quicker
* On calculate instead of individually saving each student model they
  are now bulk updated
* As well as this, only when grades are updated or delete or added for a
  given student are they flagged for re-calculation. This means the
first calculation is slightly slow. But all future loads of the page are
extremelyt fast if no chanegs are made (no calculation) or fast because
only a small number of students need recalculating